Setup SerilogRegistration in other file
